intel_realsense/requirements.txt
pyrealsense2
opencv-python
numpy
pillow
